Silence regression that surfaced after I disbled the coping of local strings in chpl__initCopy(a:string);.
This particular test depended on there being a local copy of the wide string
being passed as an argument to the extern function.  

However, when compiling no-local, we need to do this explicitly, not by
accident, so the fact that it happened to (partially) work before is no better
than it not having worked at all (and arguably worse).

In any case trying to repair this before David's reference counted strings
change goes in would be a waste of time, because that would undoubtedly break
this test anyway: extra work is required to make it work correctly.

Specifically, when we see a wide string being passed to an extern routine, we
must make a local copy of the data contained in that string, and pass a pointer
to the char[] (if passed by value).  If passed by ref, inout or out, we must
pass a pointer to the char* pointer in the struct which contains the string.  If
the variable to be updated is represented by a wide pointer, we must do a get
(including a deep copy of the character data) for the ref or inout cases.  Pass
a pointer to the char* field in the string struct (i.e. the pointer has type
char**) and then perform a put (including a deep copy) after the exter proc has
been called.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21360 3a8e244f-b0f2-452b-bcba-4c88e055c3ca