Destroy more variables as they leave scope / fix RecordDestructor1 regression
[reviewed by xyz; motivation and detective work by hilde]

A long-standing regression for --no-local/GASNet testing has been
test/classes/figueroa/RecordDestructor1.chpl in which a global record's
destructor was not being called as it is in --local compilations.  I don't
have any previous recollection of this issue, but was motivated to look at it
after yesterday's meeting, particularly after I'd been asserting that we do
call record constructors the day prior (again, true in --local compilations).
In his hierarchical locales work, Tom had tracked the source of the issue
down to the line that was removing the autodestroy call for the record (which
can be seen in the diff to parallel.cpp in this commit), but had not had a
chance to go further than that.

The net effect of this change is to tighten up the conditional that was
removing autodestroys in order to limit the removal to sync variables rather
than *all* autodestroys.  This is still not perfect -- it means we're leaking
sync variables -- but it's a step in the right direction since we now are
calling destructors in --no-local executions that we were not before.

The mechanism I used to tighten up the conditional was to add an additional
pragma to sync/single var autodestroy calls to mark them as such --
FLAG_AUTO_DESTROY_FN_SYNC and branch based on that rather than simply
FLAG_AUTO_DESTROY_FN.  If other types needed to be special-cased this way,
the flag should be renamed to something like
FLAG_AUTO_DESTROY_FN_MAY_PERSIST...

Why do sync variables need to be kept around?  As far as I can tell (noted in
the source code), it's because in the presence of tasks, they are not moved
to the heap like other value types; yet they are not reference counted like
other reference types.  So closing this gap remains future work.  I
considered filing a future against it, but decided that since it will be
reported as a memory leak (which we'll be trying to push to 0 in the coming
year), and is a pre-existing condition, the current testing covers it fine.
This topic is also likely to come up naturally as we continue to wrestle with
task intent semantics (and their implication for the lifetime of by-reference
variables).

The other thing I didn't get a clear answer to is what it is about --no-local
compilation that causes this autodestroy to be removed when it isn't under
--local.  There's no parallelism/on-clause within the user code that would
cause the behavior to be different, suggesting that it's something about
globals; perhaps the fact that they exist when parallelism used to set up the
program fires?  For awhile, I thought that the conditional in question was
only reached in --no-local compilation paths, but I found the cases that
failed for sync vars under --local testing, so that didn't turn out to be
correct.

I added a few variations of the original RecordDestructor test that
I was curious about as part of this investigation -- one compares
module-scope variables to main-scope variables; the second looks at
the impact of begins. IIRC, both worked prior to this change (and
still do).

So, not completely satisfying, but a step in a positive direction as a
background activity...




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21362 3a8e244f-b0f2-452b-bcba-4c88e055c3ca