Duplicate recent tasks=fifo hierloc changes for tasks=massivethreads.
[Reviewed by Tom.]

This adjusts massivethreads tasking for locale-aware memory allocation
in the same way that r21331 and r21345 did so for fifo tasking.

Note that massivethreads in hierloc does not work for all tests.  It
fails for release/examples/hello4-datapar-dist.chpl, for example.  That
test also fails in trunk, however, and in the same way: a timeout.  It
also fails with hello6-taskpar-dist.chpl, but that appears to be due to
a problem with r21357, which has been reported to its author.  So for
now, I'm committing this, and hoping that the cause or causes of the
failures are other than in this particular commit..

The specific file changes are as follows:

tasks-muxed.c
  - In both the task launch wrapper descriptor and the thread local data
    descriptor, replace the serial state, locale, and "here" pointer
    with a chpl_task_private_data_t, which contains the same information
    and more.
  - Add GET_STATE and SET_STATE macros to get and set all of the task
    private data at once, and use these to implement SAVE_STATE and
    RESTORE_STATE.  This may actually be overkill.  We may only need to
    save and restore a subset of the task private data, like the serial
    state, local ID, and "here" pointer.  But it's safest to save and
    restore all of it.
  - Get rid of {GET,SET}_{HERE,LOCALE}.  They are no longer needed.
  - Set task private data appropriately when the initial task descriptor
    is created in chpl_task_init(), supporting tasking operations prior
    to the point at which the user main becomes a real task.
  - Set vacuous task private information in the comm (polling) task.
  - Add an implementation of chpl_task_getPrivateData() which can work
    when called before tasking is initialized.
  - Change the order of the chpl_task_{get,set}{Serial,LocalID,Here}()
    declarations to match the order of the members themselves, within
    the task private data.
  - Implement chpl_task_{get,set}{Serial,LocalID,Here}() in terms of
    chpl_task_getPrivateData().


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21363 3a8e244f-b0f2-452b-bcba-4c88e055c3ca