Get error on destruction of records working again.
Reworked the associated semantic check to rely on a property of the destructor
function (i.e. that it has the FLAG_DESTRUCTOR flag) rather than that it calls
specifically PRIM_CHPL_FREE (which it doesn't anymore).
  compiler/passes/semanticChecks.cpp

Removed:
  test/types/records/sungeun/destructor2.future
  test/types/records/sungeun/destructor2.bad

Replaced object with opaque as declared type of the pointer passed to
chpl_memhook_free_pre (for consistency with the implementations of
chpl_here_free(s:sync) and chpl_here_free(s:single).
  modules/internal/ChapelLocale.chpl

Updated to reflect this change:
  CHECKLIST-trunk-merge.txt


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21367 3a8e244f-b0f2-452b-bcba-4c88e055c3ca