Fix problems building runtime with qthreads and RE2
There were some problems to do with getting qthreads support working from C++
(since the Chapel/RE2 integration is written in C++ and needs access to e.g.
locks which are defined by the tasking layer).

The first problem was that qthread's include directory was ending up in
RUNTIME_CFLAGS (and since RE2 support builds with C++ it would use
RUNTIME_CXXFLAGS). This patch changes it to update RUNTIME_INCLS (and also
makes the same change for Massivethreads since it had the same problem).

The second problem was that in order to link properly, the RE2 support includes
Chapel stuff in an extern "C" { } block.  Then the qthreads.h header cleverly
detects a C++ compiler and then includes some C++ stuff.... which causes errors
if it's within the extern "C" block. So, a patch to qthreads adds extern "C++"
before including the C++ aspects of qthreads.

Reviewed by Brad. The qthreads team is aware of and plans to include the
change to qthread.h.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21375 3a8e244f-b0f2-452b-bcba-4c88e055c3ca