Get rid of test_nested_var_iterator2.future.
Copy propagation functions incorrectly in the presence of aliases introduced by
var iterators, so the bug that was uncovered by locale-aware memory work remains
in place.  A simple solution would be to black-list any base variable whose
address is taken, up to the point the associated reference variable goes out of
scope or is reassigned.  

A more sophisticated solution would be to perform reference tracking.  However,
since it was not possible to easily reproduce this error using straight code,
perhaps reference propagation is not common enough in the AST to make this more
detailed approach worthwhile.

Removed.
  test/functions/deitz/nested/test_nested_var_iterator2.future

Added a new "allocator" flag to chpl_here_alloc().
  modules/internal/ChapelLocale.chpl
  compiler/include/flags_list.h

Modified scalarReplace to pay attention to the new flag.
  compiler/optimizations/scalarReplace.cpp

Added a comment.
  compiler/optimizations/copyPropagation.cpp


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21377 3a8e244f-b0f2-452b-bcba-4c88e055c3ca