Enable packed wide pointers and clean up code generation for wide pointers
[Reviewed by Tom and Greg]

Previous to this patch, Chapel used a variety of wide pointer formats:
{int(32), something*}, {int(64), something*}, {{int(32), int(32)}, something*}.
In order to build LLVM global pointer optimizations, I needed Chapel to produce
64-bit wide pointers. 64-bits is probably enough for all the memory in even a
very large machine for the forseeable future (since 64 bits can able to address
16 exbibytes). However, we need to be able to get a node number into the wide
pointers. So, we pack the wide pointers. (I call them 'packed wide pointers'
here but probably 'packed global pointers' makes more sense).

This patch does not include updates to wide pointer type code generation for
C or LLVM backends; you can expect an LLVM version using packed pointers as
part of patches I have yet to commit that implement LLVM global pointer
optimization.

First, there is now a new chplenv configuration variable called
CHPL_WIDE_POINTERS. This variable defaults to 'struct' - meaning the current
state of affairs - but can be set to 'node#' e.g. node16. That means that the
wide pointer format should consist of 16 bits of node number and the rest (more
or less) for the local pointer address. I say more or less because the current
implementation sets the high bit on all packed wide pointers and checks that it
is set when operating on one. As a result, it would be an error to:
 - dereference a global pointer directly (on linux x86-64, addresses with the
   high bit set are in kernel space)
 - treat a local pointer, other than NULL, as a wide pointer (since the
   functions to extract parts of the pointer check for the high bit)

Now, in order to implement this feature, besides introducing a new
configuration variable, I have also moved the responsibility for getting the
fields from wide pointers (ie node and address) from entirely a Chapel code
generation concern to one that uses runtime support. In particular, the new
file runtime/include/chpl-wide-ptr-fns.h contains functions for creating wide
pointers and extracting the relevant pieces. In order for these functions to
exist, we needed to define wide-pointer-to-void in chpltypes.h (we define
wide_ptr_t and ptr_wide_ptr_t). Likewise, in order to express the locale
portion of a pointer in these C functions, the definition of chpl_localeID_t
now also resides in chpltypes.h. Even if a future version of the Chapel
compiler needs to generate these types, it should produce the relevant bits
of chpltypes.h. Having them available in the runtime makes it much easier to
write generic operations on wide pointers.

The configuration variable for CHPL_WIDE_POINTERS drives some #defines that are
passed along when building the runtime or when building Chapel programs. The
chplenv/widePointers script accepts a --defines option to emit the appropriate
thing (such as -DCHPL_WIDE_POINTER_STRUCT by default).

The new wide-ptr functions are implemented with #ifdefs so that if the pointer
format is 'struct' they do the right thing, and if the pointer format is e.g.
'node16' they handle the packing and unpacking. Note that we always pass wide
pointers into these functions by pointer-to-wide-pointer. That way, it is easy
to use a cast to go from a typed wide pointer to the wide-point-to-void that
the functions understand.

I have updated the code generation routines in compiler/AST/expr.cpp, such as
codegenRaddr, to use these functions. Also, there are now separate routines for
extracting the address from a wide string since sometimes we need to get the
address of a wide string as an lvalue (e.g. passing a wide string by reference
to an extern C routine). It would not be possible to get an lvalue pointing to
the address portion of a packed wide pointer. I believe that these changes also
solve the problem Tom was fixing in r21286 but in a cleaner manner. I did not
want to confuse things with extra substitutions so I #define'd createTempRef to
be createTempVar (it could be found and replaced in a later patch). I believe
that Tom will agree that the function should be called createTempVar...

I have also updated globals registration and broadcast.  In particular, the
globals registry used to just store pointers to the address portion of the wide
pointers, which is no longer possible if we have packed wide pointers. So,
instead the registry now stores pointers to the entire wide pointer, and it is
these entire wide pointer values that are overwritten with the values from
locale 0. (In comm-gasnet.c this change mostly has the flavor of replacing
sizeof(void*) with sizeof(wide_ptr_t)). It is likely that this change will
require any comm layer other than GASNet to change.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21380 3a8e244f-b0f2-452b-bcba-4c88e055c3ca