Move deep copy of wide strings into a codegen call and generate that one case specially.
This patch fixes the passing of wide strings between nodes following the removal
of the string_copy call from chpl__initCopy(s:string).  That is, initCopy on a
local string just returns the same string -- effectively copying the char* and
not the underlying byte array.

Of course, this does not work across the network, becuase the pointer to a byte
array on a remote node has no meaning on the local node.  So (only) in the case
of wide pointers, it is necessary to perform a deep copy.

The decision as to whether a copy should be shallow or deep should be controlled
by the implementation, so this is really a one-off implementation for the
strange wide_string data type (that does not even follow standard wide pointer
format conventions).  

Representation of strings as reference-counted records, and a more complete
implementation of memory management should permit us to remove this
special-purpose code.  For now, it is set off well enough since it appears in
only one place in the codegen code and is commented appropriately.

[tested with: gasnet.fifo, test/release/examples/ only]


Add a codegen utility function to perform the steps needed for a wide string get.
  runtime/include/chpl-gen-includes.h

Add a special case in codegenAssign, to call chpl_gen_comm_wide_string_get()
instead of chpl_gen_comm_get() when the RHS is of type wideStringType.
  compiler/AST/expr.cpp


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21388 3a8e244f-b0f2-452b-bcba-4c88e055c3ca