A better implementation of chpl_gen_comm_wide_string_get ...
... because it also works with local builds.  Local builds do not define the
chpl_chpl__wide_chpl_string_s struct, so routines which access fields in it have
to go in chpltypes.c (where it's always defined, even in local builds).

Moved the definition of chpl_gen_comm_wide_string_get from chpl-gen-includes.h
to chpltypes.c, turning it from a static inline to a regular function.  Added a
declaration for it in chpl-comm.h.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21389 3a8e244f-b0f2-452b-bcba-4c88e055c3ca