Add support for standard ops on tuple/scalar pairs
[reviewed by diten]

This patch adds overloads of standard operators on pairs of homogenous
tuples and scalars of matching element type by inserting them into the
internal ChapelTuple.chpl module.  The supported operators are: +, -,
*, /, %, **, &, |, ^, <<, and >>.  The implementation essentially
applies the scalar operator to each tuple element returning a tuple
result.  I shied away from equality-oriented operators due to concerns
that it such cases may represent programmer error rather than an
intention promote an equality test across the tuple components.  This
support could be reconsidered later or made available by a standard
module if desired but considered to be a dangerous default.

We have supported such overloads in the past and backed them out.
Browsing the old logs, the rationale for removing the support most
recently (r11766) was due to the implementation approach taken rather
than the lack of utility of the functions.  In particular, the
operators were supported using signatures as follows (excuse the old
syntax):

        pragma "inline" def +(a: _tuple, b) { ... }
        pragma "inline" def +(a, b: _tuple) { ... }

and the issue was that the completely generic non-tuple argument cast
too wide of a net, catching things that it wasn't intended to.  For
example, adding a tuple to an array of tuples fell into this overload
rather than promoting the +(tuple,tuple) overload.

The commit also rationalizes that users who want this can overload it
themselves which is true, but in practice, I find that these are very
natural operators to desire when doing tuple work (like storing
velocities and positions as tuples and shifting or scaling them, as in
n-body codes), and an annoyance to implement.

The main downside of supporting these operators is that it prevents
users from writing their own overloads with the same signatures, if
they desire.  My proposed solution to this would be to move these
overloaded operators, and those on tuple-tuple pairs, into an
auto-included standard module, and adding the capability we've
discussed in the past to override the auto-use for a particular
file/scope.

The overloads that I wanted to use to implement these operators
would have ideally been something like:

        inline proc +(a:?*?t, b: t)

That is, "accept a homogenous tuple of 't's of any size and a second
argument that coerces to type 't'.  However, the formal type query
syntax as implemented isn't rich enough to do this yet.  There's also
the question of how the reverse case +(scalar, tuple) would be written and
how symmetric it should be in terms of coercible types.

Instead, I've used the following signatures:

       inline proc +(x:_tuple, y: x(1).type) where isHomogeneousTuple(x) ...

and:


        inline proc +(x:?t, y: _tuple) where isHomogeneousTuple(y) &&
                                             t: (y(1).type)  ...

where the t: (y(1).type) is intended to express "t is coercible to
the type of the tuple elements" but doesn't quite seem to work
that way in the implementation -- it's acting more like equality
(TODO: why is this?)


One other open question is whether these operators should be extended
to heterogeneous tuples as well (as the tuple-tuple forms are).  I've
chosen not to here, primarily for simplicity and because it supports
the cases that I run into commonly.  I also think there's an argument
to be made that it's hard to know how to characterize what kind of
scalar argument can be combined with a heterogeneous tuple (which is
what the previous implementation was effectively trying to do and what
got it into trouble).  I can also imagine that we might want to
re-think whether the tuple-tuple ops should only be supported (at
least by default) for homogenous tuples down the road (for simplicity,
to avoid surprises), though I'm not championing that here.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21393 3a8e244f-b0f2-452b-bcba-4c88e055c3ca