Tests for new intents (and old)
This commit adds a pair of tests that check the new intents as
applied to ints and strings (as well as the old intents) and
makes sure that const-ness is verified when it ought to be.
I've also manually checked the generated code to see that things
look consistent, but that's not tested in any automated way.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21413 3a8e244f-b0f2-452b-bcba-4c88e055c3ca