Moving existing intents test into new intents subdirectory
I had meant to do this as housecleaning yesterday, having just created
the directory.  Turns out I should've from a regression standpoint --
the presence of an intents.chpl test and intents/ directory caused the
test system to croak since it uses -o intents for that test.  I'll
work on filing a future, if not bug, for this behavior, but in the
meantime, this should fix last night's specific failure.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21416 3a8e244f-b0f2-452b-bcba-4c88e055c3ca