Subject: Adding future test for assertion error in callInfo.cpp when using a void* in an extern #include
Description: (Paraphrasing Brad here..) 
When using using an external function that returns a void* we get an
assertion error. The problem appears to be being created in externCResolve.cpp when
if the pointee is a void*, we do:
   
   pointee = new DefExpr(new ArgSymbol(INTENT_BLANK, astr("void", istr(query_uid++)), dtAny));

We then do a return of:

  new CallExpr(new UnresolvedSymExpr("_ddata"), new CallExpr(PRIM_ACTUALS_LIST, pointee));

Which makes it appear as though a DefExpr of a formal argument is being used as an
actual argument. Later on in callInfo.cpp we fail an assertion based on this issue
(argument to a function is something other than a SymExpr()). 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21425 3a8e244f-b0f2-452b-bcba-4c88e055c3ca