In the compiler, outline task functions in a new createTaskFunctions pass right before resolution, instead of during the parallel pass.
[reviewed by Sung]

This is a step towards the new "task intent semantics" aka "capture
for parallelism". This commit does not change the semantics, but
switches the implementation of task functions that we will rely on for
this change, which is upcoming.

"outlining" means replacing a BlockStmt for begin/on/cobegin/coforall with
a function call to a "task function".  The latter is a standalone function
whose body is merely that former BlockStmt (plus perhaps some plumbing).

One consequence of outlining before lowerIterators is that a task function
n may contain 'yield' statements (primitive CallExprs) that logically
belong to the iterator that invokes this task function. Quite a bit of
changes in lowerIterators address this. Basically, we want to handle
the outlined task function in the same manner as we used handle its
corresponding task-parallel or 'on' block.

While this new implementation adds complexity to the lowerIterators pass,
why did we choose to stick with it?

- We need to outline task functions sooner or later in compilation, anyway.

- Doing it this way allows us to rely on the existing code for
  (a) argument passing and (b) constness checking.

- The alternative of implementing "capture for parallelism" would be
  to create explicit shadow copies that capture the values.
  But that would increase the amount of copying in the AST and
  currently in the generated code.


OVERVIEW OF CHANGES

* In a few places where the AST of an iterator is traversed, I added a
recursive traversal of the task functions that it calls, if any.
This essentially preserves the set of traversed AST nodes in the presence
of task function outlining. Mostly this is needed to handle yields.

In some cases I outlined processing into recursive functions and put in
additional arguments to make it work. Code organization-wise, I tried
to reduce the amount of diffs that this introduces (within reason,
hopefully). Here are most of the functions that got affected:

  nonLeaderParCheck
  computeRecursiveIteratorSet
  fragmentLocalBlocks
  convertYieldsAndReturns
  fnContainsOn
  expandIteratorInline
  rebuildIterator

* I am not convinced I have covered all the cases where the above
is needed. But reasoning about it is hard, so instead I am choosing
to wait until some test exposes the missing case.

These cases, for example, are suspects:
- the code within the test
    if (iteratorFn->hasFlag(FLAG_ITERATOR_WITH_ON))
  in bundleLoopBodyFnArgsForIteratorFnCall()
- should the iterator get FLAG_ITERATOR_WITH_ON if the 'on'
  is within a task function it calls?
- countEnclosingLocalBlocks
- the case in inlineSingleYieldIterator() where I put INT_FATAL()

* I chose to flatten the newly created task function 'fcopy' in
expandBodyForIteratorInline() right away - so that it is not
iterated over during later AST traversals that may happen
in this pass. An extra benefit of that is added clarity
of the AST, useful when debugging lowerIterators.

* BTW the original of 'fcopy' (i.e. 'cfn') seems to be still inTree() -
as long as the iterator that it's called from is still inTree().
'cfn' might become unused; if we prune it, parallel() will have
less work. If unused, it will be pruned right after parallel, though.

* A couple of places in the compiler make the assumption that
a task function is invoked exactly once. Since task functions
are outlined and their call sites may get cloned within
lowerIterators, this assumption may not hold. I changed the
following to address that:

  remoteValueForwarding.cpp
  create_block_fn_wrapper(), bundleArgs()

* Regarding the new possibility of multiple calls to a task function:
I have observed it earlier in my work on this change, but did not
observe it one way or the other lately. So I cannot guarantee
that multiple calls *do* happen.

* localizeYieldSymbolsInTaskFun() is needed so that concurrent
task functions do not refer to the same shared return temp
that all the iterator yields are implemented to use.
Or maybe we should just run localizeReturnSymbols() earlier?

* In util/misc.cpp, I augmented
  printDevelErrorHeader()
with
  findNonTaskCaller()
because otherwise errors in begin/... blocks would be reported as in
begin_fn/etc., rather than in the enclosing source-level function,
which would be less ideal. findNonTaskCaller() needs to find just one
caller, because they all supposedly have the same line number
information. That's because they all should come from the same
location in the source code, even if replication happens.

* The above code is exercised by nestedForall2.chpl.
The order of errors reported in the compiler happened to be
one way before this change and another way with this change.
I updated .good with the order change, because I believe
the order is not significant.

* The machinery to outline the task functions is moved from parallel()
into createTaskFunctions() almost verbatim. The minor changes are:
- we do not add them to nestedFunctions (because we do not flatten
  them right away), and
- I uncommented the check info->isPrimitive(PRIM_BLOCK_PARAM_LOOP)
  because we are now before resolution.

* The new pass, createTaskFunctions, is very small, but I did not see
that it would blend into another existing pass. Also, the file
createTaskFunctions.cpp will get more code when we switch the semantics.

* The increased timeout for distributions/dm/t5a.chpl is needed due to
increased compilation time. The calls to flattenNestedFunctions()
in expandBodyForIteratorInline() collectively take a long time
in this test (but not, say, in SSCA2). The reason is that
flattenNestedFunctions() is at least o(#calls) (due to compute_call_sites()),
and the number of invocations is large: expandIteratorInline()
is invoked 186 times for this test under comm=gasnet (this count
was the same before this commit). I have a fix in the works
for this issue.

* The same test distributions/dm/t5a.chpl also used to spend a lot of
time doing INT_ASSERT(noOtherCalls(...)). noOtherCalls is o(#calls)
and rebuildIterator() is invoked 397 times for this test under
comm=gasnet (this count was the same before this commit).  The assert
passed during my tests, but I disabled it for this check-in to reduce
compilation time in such extreme cases.


OTHER NOTES

* On collectMySymExprs(): it is analogous in the desired result to
collectMyCallExprs() invoked with the equal first and last args.
However the latter does more work because it descends into nested symbols.
Which makes collectMyCallExprs() more general, but that extra generality
is not used. So we prefer the new implementation style.

If it is desired to have the first 'me' argument be a BaseAST*,
not necessarily a Symbol*, either of the following functions can be added:

void collectMySymExprs(BaseAST* me, Vec<SymExpr*>& symExprs) {
  // skip the isSymbol(ast) check in collectMySymExprsHelp()
  AST_CHILDREN_CALL(me, collectMySymExprsHelp, symExprs);
  if (SymExpr* se = toSymExpr(me))
    symExprs.add(se);
}
void collectMySymExprs(BaseAST* me, Vec<SymExpr*>& symExprs) {
  if (Symbol* meSym = toSymbol(me))
    collectMySymExprs(meSym, symExprs);
  else
    collectMySymExprsHelp(me, symExprs);
}

The STL version (added by Sung's request) is currently unused.

* The lowerIterators pass ignores the special meaning of task
functions (which is their association with new tasks).
lowerIterators (mostly? always?) treats them as usual functions.
One exception here is yields in task functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca