Moved "support for user-level replicated variables" from modules/dists/ReplicatedDist.chpl to the newly-added modules/standard/UtilReplicatedVar.chpl.
[reviewed by Tom]

Correspondingly, added 'use UtilReplicatedVar;' to the tests that needed it.

The reason is that the code being moved is not needed to use ReplicatedDist
per se. However, since it was in the same module, using ReplicatedDist
also engaged that code. Which creates one ReplicatedDist domain map
and a corresponding domain. Both of those are privatized and so add
overhead at startup. This commit will remove this overhead when
the feature is not used.

No code changes here, except small changes in comments.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21429 3a8e244f-b0f2-452b-bcba-4c88e055c3ca