Minor cleanups and a follow up to r21426.
[Parts reviewed by David; parts are trivial.]


* baseAST.cpp - inserted the missing "ExternBlockStmt" in astTagName.


* expr.cpp - strengthened CallExpr::verify() based on the current invariants.

I did not investigate why this check succeeds:
  toBlockStmt(parentExpr)->blockInfo != this

I just disabled it to avoid failures in most tests.


* view.cpp - added a space that was missing in 'nview' printouts of SymExpr.


* map.h - removed class PointerHashFns, because (a) it is not used and
(b) we should not use it because it results in nondeterministic behavior.
Instead something like this should be used:

 class AST_HashFns : public HashFns<const BaseAST *> {
  public:
   static unsigned int hash(const BaseAST *s) { return (unsigned int)(s->id); }
   static int equal(const BaseAST *a, const BaseAST *b) { return a == b; }
 };


* passes.h - fixed the spelling of _PASSES_H_


* parallel.cpp - renamed 'nestedFunctions' -> 'taskFunctions',
to reflect the purpose of this variable.


* lowerIterators.cpp:

- Replaced a couple of collect_asts with collectCallExpr, in those
cases where only CallExprs are of interest.

- Invoked clear() on taskFunInRecursiveIteratorSet, iteratorFnMap and
loopBodyFnArgsSuppliedMap at the end of the pass, to release some memory,
assuming this is useful.

- Removed iteratorWithOnStatementSet - it is unused.


* wrappers.cpp - moved the DefExpr of 'leaderIndex' to the enclosing BlockStmt,
i.e. up one level of nesting from where it used to go. That way the uses of
'leaderIndex' are properly scoped.

This relies on buildForLoopStmt wrapping 'body' in two levels of BlockStmts,
the outer of which gets assigned into 'loop.
If we placed that DefExpr in 'loop', it would be outside a loop. Since
that is unnecessary, it is tidier to place the DefExpr in the loop.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca