Un-futurized, (re-)enabled and beefed up testing of users/vass/crash3tuple.chpl.
Following Brad's lead, I double-checked my diagnosis in r21434 and
found that it was incorrect. It appeared that, until last night, this
test was still failing on the same INT_ASSERT as it was originally. So
it's most likely that something in the compiler got fixed, and this
test can happily compile and execute.

For *.good, I simply captured the current test output.
I had to change assertions to writelns in the test to make
the output more deterministic. BTW those assertions are not
triggered when running with --dataParTasksPerLocale=4.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21438 3a8e244f-b0f2-452b-bcba-4c88e055c3ca