Use the system's default number of CPUs per compute unit instead of the largest number on any node.
[Reviewed by Sung.]

This adjusts the command line we give to aprun when using the pbs-aprun
and aprun launchers on Cray Inc. systems.

We've been using cnselect to determine the number of CPUs per compute
unit on various nodes, and taking the maximum of these as the value we
give to aprun via its -j option.  But we also use cnselect to obtain the
number of cores per node on various nodes, and take the maximum of these
as the value we give to aprun via its -d option.  Unfortunately this
breaks when the maximum numbers of cores per node and CPUs per compute
unit don't occur on the same system nodes.  So instead, by default give
0 (zero) as the value for the aprun -j option, which tells it to use
whatever is appropriate for the nodes that have been selected.  This can
still be overridden using the CHPL_LAUNCHER_CPUS_PER_CU environment
variable.

While here, I also changed CPUsPerNode to CPUsPerCU in all variable and
function names, to reflect the fact that it's the number of CPUs per
compute unit we're talking about, not the number of CPUs per node.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21441 3a8e244f-b0f2-452b-bcba-4c88e055c3ca