Reduce memory leaks due to string copies. [reviewed by Vass]
This patch is a temporary measure to reduce local string copies and thereby
reduce overall memory leaked.  Memleaks testing showed an overall reduction from
3.5Mb to 95kb leaked.  It also performs remote string copies correctly, fixing
the long-standing no-local bugs on freadIntFailed and fwriteIntFailed.
This fix is not a complete solution, because shared strings cannot be reclaimed
without reference counting or garbage collection, and neither of those is
currently implemented.  All local string data is leaked consistently.

The important change provided by the patch is to *not* perform string copies in
=(s:string) (i.e. string assignment), nor in chpl__initCopy(s:string).  To avoid
double-deletion errors, the destructor is never called for a string variable or
field.

To perform wide string gets correctly, it was necessary to add a deep copy in
that special case.  Previously, the string_copy in (narrow) assignment was a
placeholder for wide_string_copy when the string was widened.  After taking
string_copy out of assignment, we had to add it back in as part of a wide string
get.

The solution for the freadIntFailed bug resulted from having to get the size
field of a wide string correct in more places.  Since local strings were
formerly copied quite a bit, chances were high that a wide string would be
copied locally before being used.  In this case, the length of the string would
be ignored, and CLIB str??? functions expecting a terminating NUL would read the
right number of characters.  In contrast, when a wide string is read remotely,
the size field has to be correct or the wrong number of characters is
transferred.

When a local string is simply widened, the chpl_string_widen() function sets the
size field correctly.  The situation that was causing problems was when the
character data in a wide string was getting overwritten through an extern call,
but the size field was not also getting updated (the extern call expects a
narrow string).  My solution to this was to add a new primitive
"string_normalize", and call this to fix up the size field after such extern
calls were made.

[tests: standard no-local; linux64.gnu.gasnet.fifo.pthreads examples,multilocale,distributions]


DETAILS
-------

  test/memory/shannon/memmaxExceedsLineno2.chpl
This test had to modified to elicit the expected behavior.  Before, the
assignment in the initialization of string s was sufficient to cause a copy to
be made.  That, in turn, would trigger the desired error message.  Since the
patch does not perform a string copy as part of assignment, it is necessary to
perform another operation to allocate more memory and cause the intended
overflow to occur.  
In this case, I used string concatenation (prepending a zero) to cause more
local string data to be allocated. 

  runtime/include/chplio.h
Used chpl_string instead of char* as the return type of chpl_refToString() and chpl_wideRefToString().

  runtime/include/chpl-comm.h
Updated some of the packed wide pointer changes, to use consistently use
chpl_localeID_t on the Chapel side and c_locale_t on the runtime side.

  runtime/src/chpltypes.c
Added the utility chpl_gen_comm_wide_string_get() to perform the necessary deep
copy when a wide string is read.

  runtime/include/chpl-comm-compiler-macros.h
Factored chpl_test_local() to use chpl_is_node_local(), and renamed it as
chpl_check_local() for consistency with with chpl_check_nil().

  runtime/src/chplcast.c
  runtime/src/chplio.c
Replaced chpl_glom_strings(1, <foo>) with string_copy(<foo>): the two idioms are
equivalent, but the latter more readable.
 
  modules/standard/IO.chpl
Added calls to __primitive("string_normalize") to fix up the size field in wide
strings after the data portion of it is overwritten by an extern call.

  modules/internal/ChapelBase.chpl
Do not call string_copy in assignment or chpl__initCopy on strings.

  compiler/include/primitive.h
  compiler/AST/primitive.cpp
  compiler/AST/expr.cpp
Add a new PRIM_STRING_NORMALIZE primitive.

  compiler/resolution/callDestructors.cpp
Do not free the memory referenced by string fields.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21442 3a8e244f-b0f2-452b-bcba-4c88e055c3ca